changed backend used by opencv
changed to 'Agg'

+++ test.py
@@ -1,26 +1,25 @@
+import os
 import cv2
 import matplotlib.pyplot as plt
 
-# Set Matplotlib backend to Agg
 plt.switch_backend('agg')
 
-# Load the image
-image = cv2.imread('./spidyAtDoffice.jpg')
+img = './pic/supernatural_004.png'
+if os.path.exists(img):
+    img = cv2.imread(img)
+    gray_image = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+    histogram = cv2.calcHist([gray_image], [0], None, [256], [0, 256])
 
-# Convert the image to grayscale
-gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+    plt.plot(histogram, color='black')
+    plt.xlabel('Pixel Value')
+    plt.ylabel('Frequency')
+    plt.title('Histogram of Grayscale Image')
 
-# Calculate the histogram
-histogram = cv2.calcHist([gray_image], [0], None, [256], [0, 256])
+    # Save the plot as an image file
+    plt.savefig('histogram.png')
 
-# Plot the histogram
-plt.plot(histogram, color='black')
-plt.xlabel('Pixel Value')
-plt.ylabel('Frequency')
-plt.title('Histogram of Grayscale Image')
+    # Close the plot
+    plt.close()
+else:
+    print("file doesn't exist")
 
-# Save the plot as an image file
-plt.savefig('histogram.png')
-
-# Close the plot
-plt.close()
